Refactor espnet/utils/cli_utils.py
